pyproject.toml
[tool.poetry]
name = "DTPatterRecognition"
version = "0.1.0"
description = "Set of tools to implement patter recognition algorithms on CMS DTs"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
# You need to install also ROOT
python = ">=3.9"
numpy = ">=2.0"
pandas = ">=2.2"
matplotlib = ">=3.9"
mplhep = ">=0.3"
PyYAML = ">=5.4.1"
argparse = "*"
pyqt5 = "*"
tqdm = ">=4.6"
natsort = ">=8.4"
mplDTs = { git = "https://github.com/DanielEstrada971102/mplDTs.git", branch = "main" }

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "*"

[[tool.poetry.packages]]
include = "dtpr"

[tool.poetry.scripts]
dtpr = "dtpr.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"